Don't copy if non-existent

+++ tests/conftest.py
@@ -98,10 +98,10 @@ class SemesterDir:
         self.tmp_path_factory = tmp_path_factory
         self.capfd = capfd
 
-        # If base directory is present in cache, copy it into new
-        # directory to avoid it being cleaned up, update the path in
-        # the cache and skip the test.
-        if self.cached_base_dir():
+        # If base directory is present in cache and does exist, copy
+        # it into new directory to avoid it being cleaned up, update
+        # the path in the cache and skip the test.
+        if self.cached_base_dir() and self.cached_base_dir.exists():
             new_base_dir = self.mktemp()
             self.copy_tree(self.cached_base_dir(), str(new_base_dir))
             self.base_dir = new_base_dir
